requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
ccxt==4.2.30
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pycares==4.4.0
pycparser==2.21
python-dotenv==1.0.1
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
